backend/requirements.txt
fastapi
uvicorn
google-adk
google-generativeai
pydantic
python-dotenv
httpx
